Exception and Comment

+++ src/main/java/cn/huajiahao/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package cn.huajiahao.community.controller;
 
 import cn.huajiahao.community.dto.QuestionDTO;
-import cn.huajiahao.community.mapper.QuestionMapper;
 import cn.huajiahao.community.model.Question;
 import cn.huajiahao.community.model.User;
 import cn.huajiahao.community.service.QuestionService;
@@ -26,7 +25,7 @@ public class PublishController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/publish/{id}")
-    public  String edit(@PathVariable(name = "id" )Integer id,
+    public  String edit(@PathVariable(name = "id" )Long id,
                         Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -44,7 +43,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam ("tag") String tag,
-            @RequestParam("id")Integer id,
+            @RequestParam("id")Long id,
             HttpServletRequest request,
             Model model){
         model.addAttribute("title",title);
